+++ mrv2/lib/mrvFl/mrvLanguages.cpp
@@ -51,9 +51,9 @@ struct LanguageTable
 LanguageTable kLanguages[] = {
     {_("English"), "en.UTF-8"},
     {_("Spanish"), "es.UTF-8"},
-    {_("Italian"), "it.UTF-8"},
-    {_("Traditional Chinese"), "zh.UTF-8"},
-    {_("Chinese Simplified"), "zh-CN.UTF-8"},
+    // {_("Italian"), "it.UTF-8"},
+    // {_("Traditional Chinese"), "zh.UTF-8"},
+    // {_("Chinese Simplified"), "zh-CN.UTF-8"},
     {nullptr, nullptr}
 };